requirements.txt
python==3.7.5
decord==0.6.0
mindspore-gpu==1.6.1
ml-collections==0.1.1
numpy==1.21.5
Pillow==9.0.1
PyYAML==6.0